requirements.txt
SpeechRecognitin
PyAudio
pyttsx3
openai
python-dotenv
datetime